requirements.txt
aiogram>=3.0
aiohttp
loguru
python-dotenv
fake-useragent